djangoapi/requirements.txt
amqp==2.2.2
anyjson==0.3.3
argon2-cffi==16.3.0
asgi-redis==0.14.1
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.2.0
autobahn==17.9.2
Automat==0.6.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
bs4==0.0.1
celery==4.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cloudinary==1.8.0
constantly==15.1.0
cryptography==2.0.3
cryptoshop==2.0.1
Django==1.11.3
django-braces==1.11.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-dajaxice==0.7
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
html-parser==0.2
html5lib==0.999999999
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
kombu==4.1.0
logutils==0.3.5
lxml==3.8.0
Mako==1.0.6
MarkupSafe==1.0
mock==2.0.0
msgpack-python==0.4.8
oauth2==1.9.0.post1
olefile==0.44
pbr==3.1.1
pecan==1.2.1
pep8==1.7.0
pesapal==0.0.6
phonenumbers==8.8.4
Pillow==4.2.1
ply==3.10
pm==0.1.1
pycoin==0.77
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.3.0
python-bitcoinrpc==1.0
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.5
requests==2.18.4
requests-toolbelt==0.7.0
schedule==0.4.3
six==1.10.0
sorl-thumbnail==12.3
SQLAlchemy==1.1.12
tqdm==4.8.4
txaio==2.8.2
urllib3==1.22
vine==1.1.3
waitress==1.0.2
webencodings==0.5.1
WebOb==1.7.3
WebTest==2.0.29
zope.interface==4.4.2
